+++ CatBot/math/maths.py
@@ -6,7 +6,7 @@ Code for mathematical operations/functions to be used for CatBot's math-related 
 import logging
 import math
 from re import match as re_match
-from typing import Optional
+from typing import Optional, Union
 
 import discord
 from discord import app_commands
@@ -19,6 +19,29 @@ def is_number(string: str) -> bool:
     """
 
     return bool(re_match(r"^-?\d+(\.\d+)?$", string))
+
+
+def round_on_ndigits(
+    x: Union[int, float, complex], ndigits: Optional[int], /
+) -> Union[int, float, complex]:
+    """
+    Round `x` based on `ndigits` and return the result.
+
+    :param x: Number to round
+    :type x: Union[int, float, complex]
+    :param ndigits: Number of digits to round, or None for no rounding
+    :type ndigits: Optional[int]
+    :return: The result of the rounding
+    :rtype: Union[int, float, complex]
+    """
+
+    if ndigits is None:
+        return x
+
+    if isinstance(x, complex):
+        return complex(round(x.real, ndigits), round(x.imag, ndigits))
+
+    return round(x, ndigits)
 
 
 # pylint: disable=too-many-public-methods
@@ -56,6 +79,7 @@ class MathCog(commands.Cog, name="Math Commands"):
         """
 
         logging.info("/add x=%s y=%s invoked by %s", x, y, interaction.user)
+
         await interaction.response.send_message(f"{x} + {y} = **{x + y}**")
 
     @app_commands.command(
@@ -69,11 +93,6 @@ class MathCog(commands.Cog, name="Math Commands"):
     ) -> None:
         """
         Sum up the provided numbers.
-
-        :param interaction: Interaction instance
-        :type interaction: discord.Interaction
-        :param numbers: Numbers
-        :type numbers: str
         """
 
         logging.info("/sum numbers=%s invoked by %s", repr(numbers), interaction.user)
@@ -106,42 +125,57 @@ class MathCog(commands.Cog, name="Math Commands"):
         """
 
         logging.info("/sub x=%s y=%s invoked by %s", x, y, interaction.user)
+
         await interaction.response.send_message(f"{x} - {y} = **{x - y}**")
 
     @app_commands.command(name="mul", description="Multiply two numbers")
-    @app_commands.describe(x="First number", y="Second number")
+    @app_commands.describe(
+        x="First number",
+        y="Second number",
+        ndigits="Number of digits to round the result by",
+    )
     async def mul(
         self,
         interaction: discord.Interaction,
         x: float,
         y: float,
+        ndigits: Optional[int] = 10,
     ) -> None:
         """
         Multiply `x` and `y`.
         """
 
-        logging.info("/mul x=%s y=%s invoked by %s", x, y, interaction.user)
-        await interaction.response.send_message(f"{x} * {y} = {x * y}")
+        logging.info(
+            "/mul x=%s y=%s ndigits=%s invoked by %s", x, y, ndigits, interaction.user
+        )
+
+        result = round_on_ndigits(x + y, ndigits)
+
+        await interaction.response.send_message(f"{x} * {y} = **{result}**")
 
     @app_commands.command(
         name="prod", description="Find the product of an arbitrary amount of numbers"
     )
-    @app_commands.describe(numbers="An arbitrary amount of numbers separated by commas")
+    @app_commands.describe(
+        numbers="An arbitrary amount of numbers separated by commas",
+        ndigits="Number of digits to round the result by",
+    )
     async def prod(
         self,
         interaction: discord.Interaction,
         numbers: str,
+        ndigits: Optional[int] = 10,
     ) -> None:
         """
         Find the product of the provided numbers.
-
-        :param interaction: Interaction instance
-        :type interaction: discord.Interaction
-        :param numbers: Numbers
-        :type numbers: str
         """
 
-        logging.info("/prod numbers=%s invoked by %s", repr(numbers), interaction.user)
+        logging.info(
+            "/prod numbers=%s ndigits=%s invoked by %s",
+            repr(numbers),
+            ndigits,
+            interaction.user,
+        )
 
         numbers_list = numbers.replace(" ", "").split(
             ","
@@ -152,26 +186,39 @@ class MathCog(commands.Cog, name="Math Commands"):
                 ephemeral=True,
             )
             return
-        total = math.prod(float(num) for num in numbers_list)
+
+        result = round_on_ndigits(
+            math.prod(float(num) for num in numbers_list), ndigits
+        )
 
         await interaction.response.send_message(
-            f"{" * ".join(numbers_list)} = **{total}**"
+            f"{" * ".join(numbers_list)} = **{result}**"
         )
 
     @app_commands.command(name="div", description="Divide two numbers")
-    @app_commands.describe(x="First number", y="Second number")
+    @app_commands.describe(
+        x="First number",
+        y="Second number",
+        ndigits="Number of digits to round the result by",
+    )
     async def div(
         self,
         interaction: discord.Interaction,
         x: float,
         y: float,
+        ndigits: Optional[int] = 10,
     ) -> None:
         """
         Divide `x` and `y`.
         """
 
-        logging.info("/div x=%s y=%s invoked by %s", x, y, interaction.user)
-        await interaction.response.send_message(f"{x} / {y} = {x / y}")
+        logging.info(
+            "/div x=%s y=%s ndigits=%s invoked by %s", x, y, ndigits, interaction.user
+        )
+
+        result = round_on_ndigits(x / y, ndigits)
+
+        await interaction.response.send_message(f"{x} / {y} = **{result}**")
 
     @app_commands.command(
         name="floordiv", description="Divide two numbers and floor the result"
@@ -188,24 +235,35 @@ class MathCog(commands.Cog, name="Math Commands"):
         """
 
         logging.info("/floordiv x=%s y=%s invoked by %s", x, y, interaction.user)
-        await interaction.response.send_message(f"{x} // {y} = {x // y}")
+
+        await interaction.response.send_message(f"{x} // {y} = **{x // y}**")
 
     @app_commands.command(
         name="pow", description="Raise a number to the power of another"
     )
-    @app_commands.describe(x="Base number", y="Exponent number")
+    @app_commands.describe(
+        x="Base number",
+        y="Exponent number",
+        ndigits="Number of digits to round the result by",
+    )
     async def pow(
         self,
         interaction: discord.Interaction,
         x: float,
         y: float,
+        ndigits: Optional[int] = 10,
     ) -> None:
         """
         Raise `x` to the power of `y`.
         """
 
-        logging.info("/pow x=%s y=%s invoked by %s", x, y, interaction.user)
-        await interaction.response.send_message(f"{x} ^ {y} = {x**y}")
+        logging.info(
+            "/pow x=%s y=%s ndigits=%s invoked by %s", x, y, ndigits, interaction.user
+        )
+
+        result = round_on_ndigits(x**y, ndigits)
+
+        await interaction.response.send_message(f"{x} ^ {y} = **{result}**")
 
     @app_commands.command(
         name="mod", description="Calculate the modulus of two numbers"
@@ -222,48 +280,89 @@ class MathCog(commands.Cog, name="Math Commands"):
         """
 
         logging.info("/mod x=%s y=%s invoked by %s", x, y, interaction.user)
-        await interaction.response.send_message(f"{x} % {y} = {x % y}")
+
+        await interaction.response.send_message(f"{x} % {y} = **{x % y}**")
 
     @app_commands.command(
         name="sqrt", description="Calculate the square root of a number"
     )
-    @app_commands.describe(x="Number to calculate the square root of")
-    async def sqrt(self, interaction: discord.Interaction, x: float) -> None:
+    @app_commands.describe(
+        x="Number to calculate the square root of",
+        ndigits="Number of digits to round the result by",
+    )
+    async def sqrt(
+        self,
+        interaction: discord.Interaction,
+        x: float,
+        ndigits: Optional[int] = 10,
+    ) -> None:
         """
         Calculate the square root of `x`.
         """
 
-        logging.info("/sqrt x=%s invoked by %s", x, interaction.user)
-        await interaction.response.send_message(f"sqrt({x}) = {math.sqrt(x)}")
+        logging.info(
+            "/sqrt x=%s ndigits=%s invoked by %s", x, ndigits, interaction.user
+        )
+
+        result = round_on_ndigits(math.sqrt(x), ndigits)
+
+        await interaction.response.send_message(f"sqrt({x}) = **{result}**")
 
     @app_commands.command(
         name="cbrt", description="Calculate the cube root of a number"
     )
-    @app_commands.describe(x="Number to calculate the cube root of")
-    async def cbrt(self, interaction: discord.Interaction, x: float) -> None:
+    @app_commands.describe(
+        x="Number to calculate the cube root of",
+        ndigits="Number of digits to round the result by",
+    )
+    async def cbrt(
+        self,
+        interaction: discord.Interaction,
+        x: float,
+        ndigits: Optional[int] = 10,
+    ) -> None:
         """
         Calculate the cube root of `x`.
         """
 
-        logging.info("/cbrt x=%s invoked by %s", x, interaction.user)
-        await interaction.response.send_message(f"cbrt({x}) = {math.cbrt(x)}")
+        logging.info(
+            "/cbrt x=%s ndigits=%s invoked by %s", x, ndigits, interaction.user
+        )
+
+        result = round_on_ndigits(math.cbrt(x), ndigits)
+
+        await interaction.response.send_message(f"cbrt({x}) = **{result}**")
 
     @app_commands.command(
         name="nroot", description="Calculate the nth root of a number"
     )
-    @app_commands.describe(x="Number to calculate the nth root of", n="Root number")
+    @app_commands.describe(
+        x="Number to calculate the nth root of",
+        n="Root number",
+        ndigits="Number of digits to round the result by",
+    )
     async def nroot(
         self,
         interaction: discord.Interaction,
         x: float,
         n: float,
+        ndigits: Optional[int] = 10,
     ) -> None:
         """
         Calculate the `n`th root of `x`.
         """
 
-        logging.info("/nroot x=%s n=%s invoked by %s", x, n, interaction.user)
-        await interaction.response.send_message(f"{n}root({x}) = {x ** (1 / n)}")
+        logging.info(
+            "/nroot x=%s n=%s ndigits=%s, invoked by %s",
+            x,
+            n,
+            ndigits,
+            interaction.user,
+        )
+
+        result = round_on_ndigits(x ** (1 / n), ndigits)
+
+        await interaction.response.send_message(f"{n}-root({x}) = **{result}**")
 
     @app_commands.command(
         name="abs", description="Calculate the absolute value of a number"
@@ -275,7 +374,8 @@ class MathCog(commands.Cog, name="Math Commands"):
         """
 
         logging.info("/abs x=%s invoked by %s", x, interaction.user)
-        await interaction.response.send_message(f"abs({x}) = {abs(x)}")
+
+        await interaction.response.send_message(f"abs({x}) = **{abs(x)}**")
 
     @app_commands.command(name="ceil", description="Calculate the ceiling of a number")
     @app_commands.describe(x="Number to calculate the ceiling of")
@@ -285,7 +385,8 @@ class MathCog(commands.Cog, name="Math Commands"):
         """
 
         logging.info("/ceil x=%s invoked by %s", x, interaction.user)
-        await interaction.response.send_message(f"ceil({x}) = {math.ceil(x)}")
+
+        await interaction.response.send_message(f"ceil({x}) = **{math.ceil(x)}**")
 
     @app_commands.command(name="floor", description="Calculate the floor of a number")
     @app_commands.describe(x="Number to calculate the floor of")
@@ -295,7 +396,8 @@ class MathCog(commands.Cog, name="Math Commands"):
         """
 
         logging.info("/floor x=%s invoked by %s", x, interaction.user)
-        await interaction.response.send_message(f"floor({x}) = {math.floor(x)})")
+
+        await interaction.response.send_message(f"floor({x}) = **{math.floor(x)}**")
 
     @app_commands.command(
         name="round", description="Round a number to the specified number of digits"
@@ -311,34 +413,45 @@ class MathCog(commands.Cog, name="Math Commands"):
         logging.info(
             "/round x=%s ndigits=%s invoked by %s", x, ndigits, interaction.user
         )
+
         await interaction.response.send_message(
-            f"{x} rounded to {ndigits} decimal places: {round(x, ndigits)})"
+            f"{x} rounded to {ndigits} decimal places: **{round(x, ndigits)}**)"
         )
 
     @app_commands.command(name="log", description="Calculate the logarithm of a number")
     @app_commands.describe(
         x="Number to calculate the logarithm of",
         base="Base of the logarithm; if empty, defaults to base e (natural logarithm)",
+        ndigits="Number of digits to round the result by",
     )
     async def log(
         self,
         interaction: discord.Interaction,
         x: float,
         base: Optional[float] = None,
+        ndigits: Optional[int] = 10,
     ) -> None:
         """
         Calculate the logarithm base `base` of `x`.
         """
 
-        logging.info("/log x=%s base=%s invoked by %s", x, base, interaction.user)
+        logging.info(
+            "/log x=%s base=%s ndigits=%s invoked by %s",
+            x,
+            base,
+            ndigits,
+            interaction.user,
+        )
 
         if base is None:
-            result = math.log(x)
-            await interaction.response.send_message(f"ln({x}) = {result}")
+            result = round_on_ndigits(math.log(x), ndigits)
+
+            await interaction.response.send_message(f"ln({x}) = **{result}**")
             return
 
-        result = math.log(x, base)
-        await interaction.response.send_message(f"log{base}({x}) = {result}")
+        result = round_on_ndigits(math.log(x, base), ndigits)
+
+        await interaction.response.send_message(f"log{base}({x}) = **{result}**")
 
     @app_commands.command(
         name="gcd",
@@ -356,7 +469,8 @@ class MathCog(commands.Cog, name="Math Commands"):
         """
 
         logging.info("/gcd x=%s y=%s invoked by %s", x, y, interaction.user)
-        await interaction.response.send_message(f"gcd({x}, {y}) = {math.gcd(x, y)}")
+
+        await interaction.response.send_message(f"gcd({x}, {y}) = **{math.gcd(x, y)}**")
 
     @app_commands.command(
         name="gcd-bulk",
@@ -416,7 +530,7 @@ class MathCog(commands.Cog, name="Math Commands"):
         """
 
         logging.info("/lcm x=%s y=%s invoked by %s", x, y, interaction.user)
-        await interaction.response.send_message(f"lcm({x}, {y}) = {math.lcm(x, y)}")
+        await interaction.response.send_message(f"lcm({x}, {y}) = **{math.lcm(x, y)}**")
 
     @app_commands.command(
         name="lcm-bulk",
@@ -468,6 +582,7 @@ class MathCog(commands.Cog, name="Math Commands"):
         y1="First y-coordinate",
         x2="Second x-coordinate",
         y2="Second y-coordinate",
+        ndigits="Number of digits to round the result by",
     )
     async def distance_cartesian_2d(  # pylint: disable=too-many-arguments
         self,
@@ -476,24 +591,26 @@ class MathCog(commands.Cog, name="Math Commands"):
         y1: float,
         x2: float,
         y2: float,
+        ndigits: Optional[int] = 10,
     ) -> None:
         """
         Calculate the Cartesian distance between two points in 2D.
         """
 
         logging.info(
-            "/distance cartesian-2d x1=%s y1=%s x2=%s y2=%s invoked by %s",
+            "/distance cartesian-2d x1=%s y1=%s x2=%s y2=%s ndigits=%s invoked by %s",
             x1,
             y1,
             x2,
             y2,
+            ndigits,
             interaction.user,
         )
 
+        result = round_on_ndigits(math.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2), ndigits)
+
         await interaction.response.send_message(
-            f"distance(({x1}, {y1}), ({x2}, {y2})) = {
-                math.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2)
-            }"
+            f"distance(({x1}, {y1}), ({x2}, {y2})) = **{result}**"
         )
 
     @dist_group.command(
@@ -507,6 +624,7 @@ class MathCog(commands.Cog, name="Math Commands"):
         x2="Second x-coordinate",
         y2="Second y-coordinate",
         z2="Second z-coordinate",
+        ndigits="Number of digits to round the result by",
     )
     async def distance_cartesian_3d(  # pylint: disable=too-many-arguments
         self,
@@ -517,26 +635,30 @@ class MathCog(commands.Cog, name="Math Commands"):
         x2: float,
         y2: float,
         z2: float,
+        ndigits: Optional[int] = 10,
     ) -> None:
         """
         Calculate the Cartesian distance between two 3D points.
         """
 
         logging.info(
-            "/distance cartesian-3d x1=%s y1=%s z1=%s x2=%s y2=%s z2=%s invoked by %s",
+            "/distance cartesian-3d x1=%s y1=%s z1=%s x2=%s y2=%s z2=%s ndigits=%s invoked by %s",
             x1,
             y1,
             z1,
             x2,
             y2,
             z2,
+            ndigits,
             interaction.user,
         )
 
+        result = round_on_ndigits(
+            math.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2 + (z2 - z1) ** 2), ndigits
+        )
+
         await interaction.response.send_message(
-            f"distance(({x1}, {y1}, {z1}), ({x2}, {y2}, {z2})) = {
-                math.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2 + (z2 - z1) ** 2)
-            }"
+            f"distance(({x1}, {y1}, {z1}), ({x2}, {y2}, {z2})) = **{result}**"
         )
 
     @app_commands.command(
@@ -556,7 +678,9 @@ class MathCog(commands.Cog, name="Math Commands"):
             )
             return
 
-        await interaction.response.send_message(f"factorial({x}) = {math.factorial(x)}")
+        await interaction.response.send_message(
+            f"factorial({x}) = **{math.factorial(x)}**"
+        )
 
 
 async def setup(bot: commands.Bot):